IR mapping works

+++ mapping.h
@@ -9,11 +9,11 @@
 #define MAPPING_H_
 
 void Window ();
-void Set_robot_position(double xpos, double ypos);
+void Set_robot_position(int16_t xpos, int16_t ypos);
 void Set_robot_angle_direction(uint8_t direction);
 uint8_t Get_robot_direction();
 uint8_t Wheelshifts_to_distance(uint8_t nr_of_wheel_shifts);
-void update_robot_position(uint8_t nr_of_wheel_shifts);
+void update_robot_position(uint16_t nr_of_wheel_shifts);
 uint8_t Get_robot_tile_x();
 uint8_t Get_robot_tile_y();
 void Right_side_detectable(uint8_t IR_data);
